--- a/kzg_data_availability/README.md
+++ b/kzg_data_availability/README.md
@@ -1,0 +1,5 @@
+This implmentation uses the BLST signature library
+github.com/supranational/blst
+
+For python, the library use SWIG definitions to create the binding.
+To test the library and generate the pythong bindings, run `/bindings/python/run.me` on your BLST implementation, then copy the resultant files to [./blst/]() directory. You need to have SWIG installed from your prefered package manager.