--- a/kzg_data_availability/blst/__init__.py
+++ b/kzg_data_availability/blst/__init__.py
@@ -1,0 +1,1 @@
+from blst.blst import *