--- a/kzg_data_availability/fk20_single.py
+++ b/kzg_data_availability/fk20_single.py
@@ -1,4 +1,4 @@
-from py_ecc import optimized_bls12_381 as b
+import blst
 from fft import fft
 import kzg_proofs
 from kzg_proofs import (
@@ -10,7 +10,8 @@ from kzg_proofs import (
     get_root_of_unity,
     reverse_bit_order,
     is_power_of_two,
-    eval_poly_at
+    eval_poly_at,
+    P1_INF 
 )
 
 # FK20 Method to compute all proofs
@@ -51,7 +52,7 @@ def toeplitz_part1(x):
     root_of_unity = get_root_of_unity(len(x) * 2)
     
     # Extend x with zeros (neutral element of G1)
-    xext = x + [b.Z1] * len(x)
+    xext = x + [P1_INF.dup() for _ in range(len(x))]
 
     xext_fft = fft(xext, MODULUS, root_of_unity, inv=False)
     
@@ -68,7 +69,7 @@ def toeplitz_part2(toeplitz_coefficients, xext_fft):
     root_of_unity = get_root_of_unity(len(xext_fft))
 
     toeplitz_coefficients_fft = fft(toeplitz_coefficients, MODULUS, root_of_unity, inv=False)
-    hext_fft = [b.multiply(v, w) for v, w in zip(xext_fft, toeplitz_coefficients_fft)]
+    hext_fft = [v.dup().mult(w) for v, w in zip(xext_fft, toeplitz_coefficients_fft)]
 
     return hext_fft
 
@@ -88,7 +89,7 @@ def fk20_single(polynomial, setup):
     assert is_power_of_two(len(polynomial))
     n = len(polynomial)
     
-    x = setup[0][n - 2::-1] + [b.Z1]
+    x = setup[0][n - 2::-1] + [P1_INF.dup()]
     xext_fft = toeplitz_part1(x)
     
     toeplitz_coefficients = polynomial[-1::] + [0] * (n + 1) + polynomial[1:-1]
@@ -101,7 +102,7 @@ def fk20_single(polynomial, setup):
 
 
 # Compute all n (single) proofs according to FK20 method
-def fk20_single_data_availability_optimized(polynomial, setup):
+def fk20_single_data_availability_optimized(polynomial: list[int], setup: tuple[list[blst.P1], list[blst.P2]]) -> list[blst.P1]:
     """
     Special version of the FK20 for the situation of data availability checks:
     The upper half of the polynomial coefficients is always 0, so we do not need to extend to twice the size
@@ -116,7 +117,7 @@ def fk20_single_data_availability_optimized(polynomial, setup):
     
     # Preprocessing part -- this is independent from the polynomial coefficients and can be
     # done before the polynomial is known, it only needs to be computed once
-    x = setup[0][n - 2::-1] + [b.Z1]
+    x = setup[0][n - 2::-1] + [P1_INF]
     xext_fft = toeplitz_part1(x)
     
     toeplitz_coefficients = reduced_polynomial[-1::] + [0] * (n + 1) + reduced_polynomial[1:-1]
@@ -124,13 +125,13 @@ def fk20_single_data_availability_optimized(polynomial, setup):
     # Compute the vector h from the paper using a Toeplitz matric multiplication
     h = toeplitz_part3(toeplitz_part2(toeplitz_coefficients, xext_fft))
     
-    h = h + [b.Z1] * n
+    h = h + [P1_INF.dup() for _ in range(n)] 
 
     # The proofs are the DFT of the h vector
     return fft(h, MODULUS, get_root_of_unity(2 * n))
 
 
-def data_availabilty_using_fk20(polynomial, setup):
+def data_availabilty_using_fk20(polynomial: list[int], setup: tuple[list[blst.P1], list[blst.P2]]) -> list[blst.P1]:
     """
     Computes all the KZG proofs for data availability checks. This involves sampling on the double domain
     and reordering according to reverse bit order
@@ -145,6 +146,13 @@ def data_availabilty_using_fk20(polynomial, setup):
 
 
 if __name__ == "__main__":
+    from timer import chrono
+    generate_setup = chrono(generate_setup)
+    commit_to_poly = chrono(commit_to_poly)
+    eval_poly_at = chrono(eval_poly_at)
+    check_proof_single = chrono(check_proof_single)
+    
+    import random
     polynomial = [1, 2, 3, 4, 7, 7, 7, 7, 13, 13, 13, 13, 13, 13, 13, 13]
     n = len(polynomial)
 
@@ -158,10 +166,10 @@ if __name__ == "__main__":
 
     # Now check a random position
 
-    pos = 9
+    pos = random.randrange(0, len(polynomial))
     root_of_unity = get_root_of_unity(n * 2)
     x = pow(root_of_unity, pos, MODULUS)
     y = eval_poly_at(polynomial, x)
     
     assert check_proof_single(commitment, all_proofs[reverse_bit_order(pos, 2 * n)], x, y, setup)
-    print("Single point check passed")
+    print(f"Single point check passed (idx={pos})")