--- a/kzg_data_availability/kzg_proofs.py
+++ b/kzg_data_availability/kzg_proofs.py
@@ -1,10 +1,14 @@
 from py_ecc import optimized_bls12_381 as b
+import blst
 from fft import fft
 from multicombs import lincomb
 
 # Generatore for the field
 PRIMITIVE_ROOT = 5
-MODULUS = b.curve_order
+MODULUS = 52435875175126190479447740508185965837690552500527637822603658699938581184513 #b.curve_order
+
+P1_INF = blst.P1(bytes.fromhex("400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"))
+P2_INF = blst.P2(bytes.fromhex("400000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"))
 
 assert pow(PRIMITIVE_ROOT, (MODULUS - 1) // 2, MODULUS) != 1
 assert pow(PRIMITIVE_ROOT, MODULUS - 1, MODULUS) == 1
@@ -14,19 +18,22 @@ assert pow(PRIMITIVE_ROOT, MODULUS - 1, MODULUS) == 1
 # Helpers
 #
 #########################################################################################
+def hex(i : blst.P1):
+    bytes = i.serialize()
+    return ''.join(f'{b:02x}' for b in bytes)
 
 def is_power_of_two(x):
     return x > 0 and x & (x-1) == 0
 
 
-def generate_setup(s, size):
+def generate_setup(s :int, size :int) -> tuple[list[blst.P1], list[blst.P2]]:
     """
     # Generate trusted setup, in coefficient form.
     # For data availability we always need to compute the polynomials anyway, so it makes little sense to do things in Lagrange space
     """
-    return (
-        [b.multiply(b.G1, pow(s, i, MODULUS)) for i in range(size + 1)],
-        [b.multiply(b.G2, pow(s, i, MODULUS)) for i in range(size + 1)],
+    return(
+        [blst.G1().mult( pow(s, i, MODULUS) ) for i in range(size + 1)],
+        [blst.G2().mult( pow(s, i, MODULUS) ) for i in range(size + 1)]
     )
 
 #########################################################################################
@@ -36,7 +43,7 @@ def generate_setup(s, size):
 #########################################################################################
 
 
-def get_root_of_unity(order):
+def get_root_of_unity(order: int) -> int:
     """
     Returns a root of unity of order "order"
     """
@@ -66,7 +73,7 @@ def div(x, y):
 #
 #########################################################################################
 
-def eval_poly_at(p, x):
+def eval_poly_at(p: list[int], x: int) -> int:
     """
     Evaluate polynomial p (coefficient form) at point x
     """
@@ -156,20 +163,22 @@ def get_extended_data(polynomial):
 #
 #########################################################################################
 
-def commit_to_poly(polynomial, setup):
+BLSTADDER = lambda x,y: x.dup().add(y)
+
+def commit_to_poly(polynomial: list[int], setup: tuple[list[blst.P1], list[blst.P2]]) -> blst.P1:
     """
     Kate commitment to polynomial in coefficient form
     """
-    return lincomb(setup[0][:len(polynomial)], polynomial, b.add, b.Z1)
+    return lincomb(setup[0][:len(polynomial)], polynomial, BLSTADDER, P1_INF.dup())
 
-def compute_proof_single(polynomial, x, setup):
+def compute_proof_single(polynomial: list[int], x: int, setup: tuple[list[blst.P1], list[blst.P2]]) -> blst.P1:
     """
     Compute Kate proof for polynomial in coefficient form at position x
     """
     quotient_polynomial = div_polys(polynomial, [-x, 1])
-    return lincomb(setup[0][:len(quotient_polynomial)], quotient_polynomial, b.add, b.Z1)
+    return lincomb(setup[0][:len(quotient_polynomial)], quotient_polynomial, BLSTADDER, P1_INF.dup())
 
-def check_proof_single(commitment, proof, x, y, setup):
+def check_proof_single(commitment: blst.P1, proof: blst.P1, x: int, y: int, setup: tuple[list[blst.P1], list[blst.P2]]) -> bool:
     """
     Check a proof for a Kate commitment for an evaluation f(x) = y
     """
@@ -180,14 +189,12 @@ def check_proof_single(commitment, proof, x, y, setup):
     # e([commitment - y]^(-1), [1]) * e([proof],  [s - x]) = 1_T
     #
 
-    s_minus_x = b.add(setup[1][1], b.multiply(b.neg(b.G2), x))
-    commitment_minus_y = b.add(commitment, b.multiply(b.neg(b.G1), y))
+    smx = setup[1][1].dup().add(blst.G2().neg().mult(x))
+    cmy = commitment.dup().add(blst.G1().neg().mult(y))
 
-    pairing_check = b.pairing(b.G2, b.neg(commitment_minus_y), False)
-    pairing_check *= b.pairing(s_minus_x, proof, False)
-    pairing = b.final_exponentiate(pairing_check)
-
-    return pairing == b.FQ12.one()
+    pc = blst.PT(blst.G2().to_affine(), cmy.neg().to_affine())
+    pc.mul(blst.PT(smx.to_affine(), proof.to_affine()))
+    return pc.final_exp().is_one()
 
 #########################################################################################
 #
@@ -195,16 +202,16 @@ def check_proof_single(commitment, proof, x, y, setup):
 #
 #########################################################################################
 
-def compute_proof_multi(polynomial, x, n, setup):
+def compute_proof_multi(polynomial: list[int], x: blst.P1, n:int , setup: blst.P1) -> blst.P1:
     """
     Compute Kate proof for polynomial in coefficient form at positions x * w^y where w is
     an n-th root of unity (this is the proof for one data availability sample, which consists
     of several polynomial evaluations)
     """
     quotient_polynomial = div_polys(polynomial, [-pow(x, n, MODULUS)] + [0] * (n - 1) + [1])
-    return lincomb(setup[0][:len(quotient_polynomial)], quotient_polynomial, b.add, b.Z1)
+    return lincomb(setup[0][:len(quotient_polynomial)], quotient_polynomial, BLSTADDER, P1_INF.dup())
 
-def check_proof_multi(commitment, proof, x, ys, setup):
+def check_proof_multi(commitment: blst.P1, proof: blst.P1, x: int, ys: list[int], setup: tuple[list[blst.P1], list[blst.P2]]) ->bool:
     """
     Check a proof for a Kate commitment for an evaluation f(x w^i) = y_i
     """
@@ -223,15 +230,28 @@ def check_proof_multi(commitment, proof, x, ys, setup):
     # e([commitment - interpolation_polynomial]^(-1), [1]) * e([proof],  [s^n - x^n]) = 1_T
     #
 
-    xn_minus_yn = b.add(setup[1][n], b.multiply(b.neg(b.G2), pow(x, n, MODULUS)))
-    commitment_minus_interpolation = b.add(commitment, b.neg(lincomb(
-        setup[0][:len(interpolation_polynomial)], interpolation_polynomial, b.add, b.Z1)))
-    pairing_check = b.pairing(b.G2, b.neg(commitment_minus_interpolation), False)
-    pairing_check *= b.pairing(xn_minus_yn, proof, False)
-    pairing = b.final_exponentiate(pairing_check)
-    return pairing == b.FQ12.one()
+    #xn_minus_yn = b.add(setup[1][n], b.multiply(b.neg(b.G2), pow(x, n, MODULUS)))
+    xn_minus_yn = setup[1][n].dup().add( blst.G2().neg().mult(pow(x, n, MODULUS)) )
+    #commitment_minus_interpolation = b.add(commitment, b.neg(lincomb(
+    #    setup[0][:len(interpolation_polynomial)], interpolation_polynomial, b.add, b.Z1)))
+    commitment_minus_interpolation = commitment.dup().add( lincomb(
+        setup[0][:len(interpolation_polynomial)], interpolation_polynomial, BLSTADDER, P1_INF.dup()
+        ).neg() ) 
+
+    pc = blst.PT(blst.G2().to_affine(), commitment_minus_interpolation.neg().to_affine())
+    pc.mul(blst.PT(xn_minus_yn.to_affine(), proof.to_affine()))
+    return pc.final_exp().is_one()
 
 if __name__ == "__main__":
+    from timer import chrono
+    generate_setup = chrono(generate_setup)
+    commit_to_poly = chrono(commit_to_poly)
+    eval_poly_at = chrono(eval_poly_at)
+    check_proof_single = chrono(check_proof_single)
+    compute_proof_multi = chrono(compute_proof_multi)
+    check_proof_multi = chrono(check_proof_multi)
+
+
     polynomial = [1, 2, 3, 4, 7, 7, 7, 7, 13, 13, 13, 13, 13, 13, 13, 13]
     n = len(polynomial)